requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
scikit-learn
pandas
python-dotenv
google-generativeai
